refactor(monitor): split notifying functions into Notifier class
