+++ src/helpers/singleton/__init__.py
@@ -1,0 +1,3 @@
+from ._singleton import Singleton
+
+__all__ = ['Singleton']
