+++ src/helpers/singleton/_singleton.py
@@ -1,0 +1,18 @@
+from __future__ import annotations
+
+from typing import ClassVar
+
+
+class Singleton:
+    _singleton: ClassVar = None
+
+    def __new__(cls, *args, **kwargs):
+        if cls._singleton is None:
+            return object.__new__(cls)
+        raise RuntimeError('A singleton instance already exists, use get_instance() instead.')
+
+    @classmethod
+    def get_instance(cls) -> 'Singleton':
+        if cls._singleton is None:
+            cls._singleton = cls()  # implicitly calls __new__ then __init__
+        return cls._singleton
